- improved getting user info - improved changing user info - fixed issue with logout - improved changing password -

+++ src/UI/User/UserNotificationCell.m
@@ -36,7 +36,7 @@
 
 - (IBAction) onSwitchBtn: (UISwitch*) sender
 {
-    NSLog(@"Switch nubmer %ld", sender.tag);
+    NSLog(@"Switch nubmer %ld", (long)sender.tag);
 }
 
 @end
